desktop/src/local-ai-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
ollama==0.3.0
sentence-transformers==3.0.1
spacy==3.8.0
pydantic==2.9.2
python-dotenv==1.0.1
numpy==2.1.1
torch==2.5.1
transformers==4.46.0

# Enhanced extraction dependencies
easyocr==1.7.1
pdfplumber==0.11.0
tabula-py==2.9.0
pandas==2.2.2
pillow==10.4.0
opencv-python==4.10.0.84
pymupdf==1.24.8  # For better PDF handling

# Test dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.1
httpx>=0.24.0

# Linting
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Add LlamaIndex dependencies
llama-index==0.10.57
llama-index-llms-ollama==0.1.4
llama-index-embeddings-huggingface==0.1.2
llama-index-readers-file==0.1.9
llama-index-readers-pdf==0.1.4